test: validate all benchmark scenarios and stress runs for all DI adapters
- Successfully executed all scenarios (register, chain, asyncChain, named, override, etc) for Cherrypick, GetIt, and Riverpod
- Verified correct integration of fully generic DIAdapter & universalRegistration architecture
- Ensured type-safety in all setup/registration flows after complete refactor
- All benchmarks run and pass under stress/load params for each DI adapter
