+++ benchmark_di/lib/di_adapters/riverpod_adapter.dart
@@ -1,4 +1,5 @@
-import 'package:benchmark_di/scenarios/universal_chain_module.dart';
+import 'package:benchmark_di/scenarios/universal_binding_mode.dart';
+import 'package:benchmark_di/scenarios/universal_scenario.dart';
 import 'package:benchmark_di/scenarios/universal_service.dart';
 import 'package:riverpod/riverpod.dart' as rp;
 import 'di_adapter.dart';
