+++ benchmark_di/lib/scenarios/universal_binding_mode.dart
@@ -1,0 +1,11 @@
+/// Enum to represent the DI registration/binding mode.
+enum UniversalBindingMode {
+  /// Singleton/provider binding.
+  singletonStrategy,
+
+  /// Factory-based binding.
+  factoryStrategy,
+
+  /// Async-based binding.
+  asyncStrategy,
+}
